requirements.txt
fire>=0.1.3
regex==2017.4.5
requests==2.21.0
tqdm==4.31.1
toposort==1.5
tensorflow==1.15.0
flask==1.1.2
google-cloud-bigtable==1.5.1
google-cloud-core==1.4.3
flask-cors==3.0.9
